Pull serial-specific bits out of general TNC interface.
